food-bot-master/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
SQLAlchemy==2.0.7
asyncpg==0.29.0
